voxel/voxel-backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Config & validation
pydantic==2.6.0
pydantic-settings==2.1.0

# ML / ASR
transformers==4.37.2
torch==2.1.2
torchaudio==2.1.2
datasets==2.16.1
huggingface-hub==0.20.3
accelerate==0.26.1

# Audio processing
librosa==0.10.1
soundfile==0.12.1
noisereduce==3.0.2
webrtcvad==2.0.10
pydub==0.25.1
numpy==1.26.3

# Database / Auth
supabase==2.3.4
python-jose[cryptography]==3.3.0

# HTTP / Utils
httpx==0.25.2
python-dotenv==1.0.0
tenacity==8.2.3

# OpenAI (for LLM text reconstruction)
openai==1.30.1
